Refactor update work
