+++ tasmoadmin/pages/selfupdate.php
@@ -1,20 +1,26 @@
 <?php
 
 use TasmoAdmin\SelfUpdate;
+use TasmoAdmin\Update\UpdateChecker;
 
 require_once _APPROOT_ . "vendor/autoload.php";
 
 $mdParser = new Parsedown();
 
 $msg        = "";
-$Selfupdate = new SelfUpdate($Config);
+
+$updateChecker = new UpdateChecker(
+    $Config->read("update_channel"),
+    $Config->read("current_git_tag")
+);
 
 if (isset($_REQUEST["selfupdate"]) || isset($_GET["selfupdate"])) {
-	$updateResult = $Selfupdate->update();
+    $Selfupdate = new SelfUpdate($Config);
+	$updateResult = $Selfupdate->update($_POST['release_url'], $_POST['latest_tag']);
 	$msg          = implode("<br/>", $updateResult);
 }
 
-$newUpdate = $Selfupdate->checkForUpdate();
+$newUpdate = $updateChecker->checkForUpdate();
 
 $changelogUrl = "https://raw.githubusercontent.com/TasmoAdmin/TasmoAdmin/master/CHANGELOG.md?r=" . time();
 $ch           = curl_init();
@@ -44,12 +50,7 @@ else {
 				<?php echo __("HELP_CHANGELOG", "NAVI"); ?>
 			<?php endif; ?>
 		</h2>
-		
-		<!--	<p class='warning'>-->
-		<!--		--><?php //echo __( "SELFUPDATE_WARNING", "SELFUPDATE" ); ?>
-		<!--	</p>-->
-		<!--	<br/>-->
-		<!--	<br/>-->
+
 		<?php if (isset($msg) && $msg != ""): ?>
 			<div class="alert alert-success alert-dismissible fade show mb-5" data-dismiss="alert" role="alert">
 				<?php echo $msg; ?>
@@ -85,8 +86,7 @@ else {
 							"SELFUPDATE",
 							[
 								$Selfupdate->getCurrentTag()
-									? $Selfupdate->getCurrentTag()
-									: __(
+									?: __(
 									"UNKNOWN",
 									"SELFUPDATE"
 								),
@@ -102,11 +102,9 @@ else {
 				</div>
 				<div class='col col-12 col-sm-5'>
 					<a class='btn btn-primary w-100 btn-green'
-					   <?php if (!empty(
-					   $Selfupdate->getLatestTag()
-					   )): ?>href='https://github.com/TasmoAdmin/TasmoAdmin/releases/tag/<?php echo $Selfupdate->getLatestTag(); ?>'
+					   <?php if (!empty($newUpdate["latest_tag"])): ?>href='https://github.com/TasmoAdmin/TasmoAdmin/releases/tag/<?php echo $newUpdate["latest_tag"]; ?>'
 					   target='_blank' <?php endif; ?>>
-						<?php echo __("NEW_TAG_VERSION", "SELFUPDATE", [$Selfupdate->getLatestTag()]); ?>
+						<?php echo __("NEW_TAG_VERSION", "SELFUPDATE", [$newUpdate["latest_tag"]]); ?>
 					</a>
 				
 				</div>
@@ -115,7 +113,9 @@ else {
 				<div class='row justify-content-sm-center mt-5'>
 					<div class="col col-12 col-sm-6 col-md-3 col-lg-4 text-center">
 						<form name='selfupdateform' method='post'>
-							<button type='submit' name='selfupdate' value='selfupdate' class='btn btn-primary'>
+                            <input type="hidden" name="latest_tag" value="<?php echo $newUpdate["latest_tag"]; ?>" />
+                            <input type="hidden" name="release_url" value="<?php echo $newUpdate["release_url"]; ?>" />
+                            <button type='submit' name='selfupdate' value='selfupdate' class='btn btn-primary'>
 								<?php echo __("BTN_START_SELFUPDATE", "SELFUPDATE"); ?>
 							</button>
 						</form>
@@ -126,17 +126,6 @@ else {
 			<div class="alert alert-info fade show mb-5" role="alert">
 				<?php echo __("NO_UPDATE_FOUND", "SELFUPDATE"); ?>
 			</div>
-			<?php if (in_array($Config->read("update_channel"), ["dev", "beta", "stable"])): ?>
-				<div class='row justify-content-sm-center mt-5'>
-					<div class="col col-12 col-sm-6 text-center">
-						<form name='selfupdateform' method='post'>
-							<button type='submit' name='selfupdate' value='selfupdate' class='btn btn-secondary'>
-								<?php echo __("BTN_START_SELFUPDATE", "SELFUPDATE"); ?>
-							</button>
-						</form>
-					</div>
-				</div>
-			<?php endif; ?>
 		<?php endif; ?>
 		<hr class='my-5'>
 		<div class='changelog'>
